django_project/requirements.txt
# Django>=4.2
# mongoengine>=0.27.0
# djangorestframework>=3.15
# django-cors-headers>=4.0.0
# redis>=5.0.1
# django-redis>=5.4.0

Django>=4.2,<5.0
mongoengine>=0.27
dnspython>=2.0.0  # Required by MongoEngine for MongoDB URI support
python-dotenv>=1.0.0
redis>=5.0.0
django-redis>=5.4.0
gunicorn>=21.2.0
pillow
django-ckeditor